owner|refactor product details and update product

+++ owner/src/main/java/org/the_chance/honeymart/ui/features/category/CategoriesViewModel.kt
@@ -457,7 +457,7 @@ class CategoriesViewModel @Inject constructor(
                 updateProductDetailsUseCase(
                     id = product.id,
                     name = product.productName,
-                    price = product.productPrice,
+                    price = product.productPrice.toDouble(),
                     description = product.productDescription,
                 )
             },
